+++ apps/client/src/common/components/link/app-link/AppLink.module.scss
@@ -1,4 +1,4 @@
-@use "../../../theme/mixins" as *;
+@use "../../../../theme/mixins" as *;
 
 .link {
   @include action-link;
