+++ apps/client/src/common/components/link/external-link/ExternalLink.module.scss
