+++ apps/client/src/features/app-settings/panel/about-panel/AboutPanel.tsx
@@ -1,4 +1,4 @@
-import ExternalLink from '../../../../common/components/external-link/ExternalLink';
+import ExternalLink from '../../../../common/components/link/external-link/ExternalLink';
 import {
   buyMeACoffeeUrl,
   discordUrl,
